+++ painttool.h
@@ -12,8 +12,8 @@ class PaintTool : public QObject
 public:
     PaintTool(Grid* grid);
     void setSize(int size) {this->mSize = size;}
-    void setHardness(float hardness) {this->mHardness = hardness;}
-    void setMinValue(float minValue) {this->mMinValue = minValue;}
+    void setHardness(double hardness) {this->mHardness = hardness;}
+    void setMinValue(double minValue) {this->mMinValue = minValue;}
     void setGrid(Grid* grid) {this->mGrid = grid;}
     PaintToolController *getController();
 public slots:
@@ -22,10 +22,10 @@ public slots:
     void setPoint(int x, int y);
     void resetPoint(int x, int y);
 private:
-    void drawPoint(int x, int y, float value);
+    void drawPoint(int x, int y, double value);
     int mSize = 1;
-    float mHardness = 1;
-    float mMinValue = -99999;
+    double mHardness = 1;
+    double mMinValue = -99999;
     Grid* mGrid = nullptr;
 };
